fix: implement authentication and session handling in ToolsPage and SubBadgeCreator components
